Move setAPIServerServiceClusterIP function to kubeapiserverexposure.go
